refactor(#3038): move turn bridge headless delivery helpers
